+++ NS_Education/Models/APIItems/Hall/GetList/Hall_GetList_Output_Row_APIItem.cs
@@ -16,7 +16,7 @@ namespace NS_Education.Models.APIItems.Hall.GetList
         public bool PrintCheckFlag { get; set; }
         public bool Invoice3Flag { get; set; }
         public int CheckType { get; set; }
-        public double BusinessTaxRate { get; set; }
+        public decimal BusinessTaxRate { get; set; }
         public int DeviceCt { get; set; }
         public int SiteCt { get; set; }
         public int PartnerItemCt { get; set; }
